fix(research): satisfy strict TS in the five probes — noUncheckedIndexedAccess, unused import, StarRing shape
`bun` runs these without typechecking, so CI's `tsc --strict` caught what my local
runs could not. All five still produce identical numbers; the probes are unchanged
in substance.

- `wset-ngfuse-is-residual-program.ts` — drop the unused `ngFromNp` import.
- `belief-fold-destructive-interference-falsifier.ts` — `lik[i]` is
  `bigint | undefined` under `noUncheckedIndexedAccess`.
- `wset-sumproduct-marginalisation.ts` — `StarRing` has no `star`/`conjugate`
  members (I invented them); explicit parameter types; guard the indexed reads
  rather than assert them, so a length mismatch throws instead of comparing
  undefined.
- `wset-doubling-transfer-table.ts` — replace the inlined basis-vector
  expression with a `basis(dim, k)` helper. It was unreadable and it was where
  the type errors lived. The output now also reports |a|, |b| and |ab| for the
  sedenion zero divisor, which is strictly better evidence: |a|=1.414,
  |b|=1.414, |ab|=0.0e+0.

Verified: `node node_modules/typescript/bin/tsc --noEmit --pretty false` rc=0,
0 errors; all five probes rc=0 with their sabotage controls still passing.

Agency-Signature-Version: 1
Agent: lumen
Agent-Runtime: claude-code
Agent-Model: claude-opus-5
Credential-Identity: AceHack
Credential-Mode: shared
Human-Review: not-implied-by-credential
Human-Review-Evidence: none
Action-Mode: human-directed
Task: none

+++ src/Core.TypeScript/research/belief-fold-destructive-interference-falsifier.ts
@@ -4,7 +4,7 @@
 // src/Core/BeliefConvergence.fs:33  ->  let observe likelihood belief = Array.map2 (*) likelihood belief
 //   ... over int64[] NON-NEGATIVE weights.
 
-const observe = (lik: bigint[], bel: bigint[]) => bel.map((b, i) => b * lik[i]);
+const observe = (lik: bigint[], bel: bigint[]) => bel.map((b, i) => b * (lik[i] as bigint));
 let seed = 24680;
 const rnd = () => { seed = (seed * 1103515245 + 12345) & 0x7fffffff; return seed / 0x7fffffff; };
 const randBelief = (n: number) => Array.from({ length: n }, () => BigInt(1 + Math.floor(rnd() * 50)));
